+++ extensions/reclaim-ai/src/hooks/useEvent.tsx
@@ -5,11 +5,10 @@ import { useMemo } from "react";
 import { Event } from "../types/event";
 import { NativePreferences } from "../types/preferences";
 import { SmartHabit } from "../types/smart-series";
-import { axiosPromiseData } from "../utils/axiosPromise";
 import { formatDisplayEventHours, formatDisplayHours } from "../utils/dates";
 import { filterMultipleOutDuplicateEvents } from "../utils/events";
 import { stripPlannerEmojis } from "../utils/string";
-import reclaimApi from "./useApi";
+import useApi from "./useApi";
 import { useCallbackSafeRef } from "./useCallbackSafeRef";
 import { ApiResponseEvents, EventActions } from "./useEvent.types";
 import { useSmartHabits } from "./useSmartHabits";
@@ -55,7 +54,7 @@ export const useEvents = ({ start, end }: { start: Date; end: Date }) => {
 };
 
 export const useEventActions = () => {
-  const { fetcher } = reclaimApi();
+  const { fetchPromise } = useApi();
   const { currentUser } = useUser();
   const { startTask, restartTask, stopTask } = useTaskActions();
   const { apiUrl } = getPreferenceValues<NativePreferences>();
@@ -80,7 +79,7 @@ export const useEventActions = () => {
   const handleStartHabit = async (id: string, title: string) => {
     try {
       await showHUD("Started Habit: " + stripPlannerEmojis(title));
-      const [habit, error] = await axiosPromiseData(fetcher(`/planner/start/habit/${id}`, { method: "POST" }));
+      const [habit, error] = await fetchPromise(`/planner/start/habit/${id}`, { method: "POST" });
       if (!habit || error) throw error;
       return habit;
     } catch (error) {
@@ -92,7 +91,7 @@ export const useEventActions = () => {
   const handleRestartHabit = async (id: string, title: string) => {
     try {
       await showHUD("Restarted Habit: " + stripPlannerEmojis(title));
-      const [habit, error] = await axiosPromiseData(fetcher(`/planner/restart/habit/${id}`, { method: "POST" }));
+      const [habit, error] = await fetchPromise(`/planner/restart/habit/${id}`, { method: "POST" });
       if (!habit || error) throw error;
       return habit;
     } catch (error) {
@@ -104,7 +103,7 @@ export const useEventActions = () => {
   const handleStopHabit = async (id: string, title: string) => {
     try {
       await showHUD("Stopped Habit: " + stripPlannerEmojis(title));
-      const [habit, error] = await axiosPromiseData(fetcher(`/planner/stop/habit/${id}`, { method: "POST" }));
+      const [habit, error] = await fetchPromise(`/planner/stop/habit/${id}`, { method: "POST" });
       if (!habit || error) throw error;
 
       return habit;
@@ -117,9 +116,7 @@ export const useEventActions = () => {
   const handleStartOrRestartSmartHabit = async (lineageId: string, title: string) => {
     try {
       await showHUD("Started Habit: " + stripPlannerEmojis(title));
-      const [habit, error] = await axiosPromiseData(
-        fetcher(`/smart-habits/planner/${lineageId}/start`, { method: "POST" })
-      );
+      const [habit, error] = await fetchPromise(`/smart-habits/planner/${lineageId}/start`, { method: "POST" });
       if (!habit || error) throw error;
 
       return habit;
@@ -132,9 +129,7 @@ export const useEventActions = () => {
   const handleStopSmartHabit = async (lineageId: string, title: string) => {
     try {
       await showHUD("Stopped Habit: " + stripPlannerEmojis(title));
-      const [habit, error] = await axiosPromiseData(
-        fetcher(`/smart-habits/planner/${lineageId}/stop`, { method: "POST" })
-      );
+      const [habit, error] = await fetchPromise(`/smart-habits/planner/${lineageId}/stop`, { method: "POST" });
       if (!habit || error) throw error;
 
       return habit;