+++ extensions/reclaim-ai/src/hooks/useSmartHabits.tsx
@@ -1,17 +1,19 @@
 import { useMemo } from "react";
 import { SmartHabit } from "../types/smart-series";
-import { fetcher } from "../utils/axiosPromise";
 import { normalize } from "../utils/objects";
 import { useSyncCachedPromise } from "./useSyncCachedPromise";
+import useApi from "./useApi";
 
 const useSmartHabits = () => {
+  const { fetcher } = useApi();
+
   const {
     error,
     isLoading,
     data: smartHabits,
   } = useSyncCachedPromise<readonly SmartHabit[]>(
     "use-smart-habits",
-    async () => (await fetcher<readonly SmartHabit[]>("/smart-habits")).data,
+    async () => await fetcher<readonly SmartHabit[]>("/smart-habits"),
     {
       wrapError: (cause) => new Error("Error while fetching smart-habits", { cause }),
     }