+++ extensions/reclaim-ai/CHANGELOG.md
@@ -1,5 +1,8 @@
 # reclaim Changelog
 
+## [Fixes] - 2024-09-03
+- Replace axios with node-fetch
+
 ## [Fixes] - 2024-08-23
 - Stopped rapid-fire calls to `GET:/api/smart-habits`
     - New hook `useSyncCachedPromise`